Upgraded from SDL2 to SDL3.

+++ JJs2DEngine/DoNotInclude/DNIData/Main/MainInternal.cpp
@@ -336,18 +336,18 @@ namespace JJs2DEngine
 		return _vertexDataMain->GetGammaValue();
 	}
 
-	float MainInternal::TranslatePositionXInPixelsToWindowSizeRatio(int32_t positionX) const
+	float MainInternal::TranslatePositionXInPixelsToWindowSizeRatio(float positionX) const
 	{
 		uint32_t windowWidth = _windowData->GetWindowWidth();
 
-		return static_cast<float>(positionX) / static_cast<float>(windowWidth);
+		return positionX / static_cast<float>(windowWidth);
 	}
 
-	float MainInternal::TranslatePositionYInPixelsToWindowSizeRatio(int32_t positionY) const
+	float MainInternal::TranslatePositionYInPixelsToWindowSizeRatio(float positionY) const
 	{
 		uint32_t windowHeight = _windowData->GetWindowHeight();
 
-		return static_cast<float>(positionY) / static_cast<float>(windowHeight);
+		return positionY / static_cast<float>(windowHeight);
 	}
 
 	InputDataListInternal& MainInternal::GetInputDataList()
