+++ JJs2DEngine/Include/Main/Main.h
@@ -44,8 +44,8 @@ namespace JJs2DEngine
 		void SetGammaValue(float newGammaValue);
 		float GetGammaValue() const;
 
-		float TranslatePositionXInPixelsToWindowSizeRatio(int32_t positionX) const;
-		float TranslatePositionYInPixelsToWindowSizeRatio(int32_t positionY) const;
+		float TranslatePositionXInPixelsToWindowSizeRatio(float positionX) const;
+		float TranslatePositionYInPixelsToWindowSizeRatio(float positionY) const;
 
 		InputDataList GetInputDataList();
 		const InputDataList GetInputDataList() const;
